refactor: remove the output stream :(
This is because Anthropic and maybe others need the structured output for context, and it's easier to do (sort of) by saving structured data. Plus I think saving structured data will be better in the long run.

I really liked the idea of having teh Output_Stream that would write to multiple, registered outputs with a single call but I have to let it go.