[#94] Manually handle scrolling for wrapped lines
Charm's bubbles/viewport (or somewhere in their stuff) has a bug (or
'feature') where it doesn't handle scrolling properly for wrapped lines.
It doesn't calculate total lines when they are wrapped; it looks only at
newlines. (https://github.com/charmbracelet/bubbles/issues/479) There
was a PR submitted for it, but was rejected because it might break
existing code and the Charm folks' philosophy has been 'leave wrapping
up as an exercise for the user'. (ie they don't want to deal with it
themselves)

So this general suggestion is to have lipgloss render it before setting
it in the viewport:
```
wrapped := lipgloss.NewStyle().Width(yourWidth).Render(yourContent)
viewport.SetContent(wrapped)
```

That's what we're doing here.

(also cleaned up comments and whatnot)