[#105] Fix response DB storage for non-tui
This was still expecting the function to return the full response in the resp variable. The new channels infra changed that. It's possible I can get that to work again, but not sure. So need to ensure the full response goes into the DB, esp. since that's the only place we're storing conversation information.